requirements.txt
requests
google-cloud-datastore
google-cloud-storage